# duplicate detection performance improvements