# Description
Jira Ticket: N/A

### Findings/changes for duplicate detection performance improvements:

- [x] separate current_user.viewable patients to avoid redundant query
- [x] do not use SQL trim on user_defined_id_statelocal because db can no longer use index
- [x] do not use SQL for casting dob, do it in ruby instead
- [ ] not sure why ruby is adding an unnecessary join on jurisdictions for the line 754 query
- [ ] can we remove jurisdiction validation?
- [ ] can we combine duplicate detection queries? what is this feature going to look like in the future? 

### Queries that are typically fired upon every iteration of the import loop:

1. Validate user provided jurisdiction path with jurisdictions in db, can't get rid of this one

```
CACHE Jurisdiction Load (0.0ms)  SELECT `jurisdictions`.* FROM `jurisdictions` WHERE `jurisdictions`.`path` = 'USA' ORDER BY `jurisdictions`.`id` ASC LIMIT 1
↳ app/controllers/import_controller.rb:249:in `validate_jurisdiction'
```

2. Calling `Patient.valid?` automatically invokes this query which validates `jurisdiction_id`. This is somewhat redundant and unnecessary, but the only way I know how to remove it is by adding `:validate => false` to `belongs_to :jurisdiction` which I'm not sure if we want to do

```
CACHE Jurisdiction Load (0.0ms)  SELECT `jurisdictions`.* FROM `jurisdictions` WHERE `jurisdictions`.`id` = 1 LIMIT 1
↳ app/controllers/import_controller.rb:91:in `block in import'
```

3. This query is fired as a result of calling `current_user.viewable_patients` in every loop, this PR includes a change which only requires this query to be called once instead of for every patient being imported

```
CACHE  (0.0ms)  SELECT `jurisdictions`.`id` FROM `jurisdictions` WHERE ((`jurisdictions`.`ancestry` LIKE '1/%' OR `jurisdictions`.`ancestry` = '1') OR `jurisdictions`.`id` = 1)
↳ app/models/jurisdiction.rb:25:in `all_patients_including_purged'
```

4. This is the first of two queries in the duplicate detection method. Calling `trim` in SQL inhibits performance severely because `user_defined_id_statelocal` is an indexed field and calling `trim` on it causes that index to be unusable. This PR includes a change to remove the use of `trim` which changes the query response time from about 8ms to 0.5ms, but we will need to run some kind of script/migration to trim all whitespaces for this field and also ensure that future whitespace does not make its way into the db

```
Patient Load (8.3ms)  SELECT `patients`.* FROM `patients` WHERE `patients`.`jurisdiction_id` IN (1, 2, 5, 3, 4, 6, 7) AND `patients`.`purged` = FALSE AND (trim(user_defined_id_statelocal) = 'EX-601120')
↳ app/models/patient.rb:756:in `duplicate_data_detection'
```

5. A redundant query was being made as a result of calling `dup_statelocal_id.present?` on top of `dup_statelocal_id.size`. This has been addressed by rewriting the code in a way in which only `dup_statelocal_id.size` is needed

```
CACHE Jurisdiction Load (0.0ms)  SELECT `jurisdictions`.* FROM `jurisdictions` WHERE `jurisdictions`.`id` = 1
↳ app/models/patient.rb:756:in `duplicate_data_detection'
```

6. This is the second of two queries in the duplicate detection method. The response time of the query seems to be ok, but I also noticed that a `LEFT OUTER JOIN` is added for no reason. I haven't done a deep dive into it yet, but I'm not sure why or how that was added to the query. Depending on the future of this duplicate detection feature, we could potentially consider combining these two (or more in the future) queries to see if that could be more performant

```
(0.5ms)  SELECT `patients`.`sex`, CAST(date_of_birth AS CHAR) FROM `patients` LEFT OUTER JOIN `jurisdictions` ON `jurisdictions`.`id` = `patients`.`jurisdiction_id` WHERE `patients`.`jurisdiction_id` IN (1, 2, 5, 3, 4, 6, 7) AND `patients`.`purged` = FALSE AND `patients`.`first_name` = 'Luana74' AND `patients`.`last_name` = 'Conn66'
↳ app/models/patient.rb:771:in `duplicate_data_detection'
```

### Additional Ideas

- refactor import controller code to allow certain queries to be executed in bulk or batches (would require a decent amount of work)
- load import page first and use AJAX requests on the fly to run the import code (more of a ux improvement but probably too complicated)

# Checklists

**Submitter:**
- [x] This PR describes why these changes were made.
- [x] This PR is into the correct branch.
- [x] **N/A** This PR includes the correct JIRA ticket reference.
- [x] Code diff has been reviewed (it **does not** contain: additional white space, not applicable code changes, debug statements, etc.)
- [x] **N/A** If UI changes have been made, Chrome Dev Tools Lighthouse accessibility test has been executed to ensure no 508 issues were introduced.
- [x] Tests are included and test edge cases
- [x] Tests have been run locally and pass (remember to update Gemfile when applicable)
- [x] Test fixtures updated and documented as necessary


@holmesie :
- [ ] Code is maintainable and reusable, reuses existing code and infrastructure where appropriate, and accomplishes the task’s purpose
- [ ] The tests appropriately test the new code, including edge cases
- [ ] You have tried to break the code
- [ ] If applicable, you have tested changes against a large database, and considered possible performance regressions


@tstrass :
- [ ] Code is maintainable and reusable, reuses existing code and infrastructure where appropriate, and accomplishes the task’s purpose
- [ ] The tests appropriately test the new code, including edge cases
- [ ] You have tried to break the code
- [ ] If applicable, you have tested changes against a large database, and considered possible performance regressions


@dubem7 :
- [ ] Code is maintainable and reusable, reuses existing code and infrastructure where appropriate, and accomplishes the task’s purpose
- [ ] The tests appropriately test the new code, including edge cases
- [ ] You have tried to break the code
- [ ] If applicable, you have tested changes against a large database, and considered possible performance regressions


@Bialogs :
- [ ] Code is maintainable and reusable, reuses existing code and infrastructure where appropriate, and accomplishes the task’s purpose
- [ ] The tests appropriately test the new code, including edge cases
- [ ] You have tried to break the code
- [ ] If applicable, you have tested changes against a large database, and considered possible performance regressions